requirements.txt
pytest==8.0.2
selene>=2.0.0rc8
selenium
allure-pytest==2.13.2
python-dotenv